[Refactor][AscendStore] Drop the gva_protocol alias module
With all consumers on the protocol functions, the backward-compatible gva_protocol.py alias (and the GVAKeyFactory name with it) is deleted. Its tests move into test_backend.py's memcache section: the exclusivity test now asserts that exposing the protocol functions, overriding the five layerwise store calls and carrying the registry marker are three views of one fact, and the byte-for-byte key format snapshots assert the module-level functions directly.
